+++ docs/source/releases.rst
@@ -26,6 +26,8 @@ Features
 Improvements
 ------------
 
+- Add support for FFTW wisdom for C++ routines.
+
 - Refactor gamma function computations.
 
 - Refactor array operation checks.